Use uint8_t for alignment in pbo_data
The pbo_data structure must be exactly 16 bytes as noted in the comment
just before it.  This is true when compiled with GCC but not with MSVC
or Clang because the latter treat a different type as a new storage
unit.  Switching to uint16_t in the middle of the pbo_data union jumps
to the next byte in the structure.  Because pbo_data ends up being
bigger than it should be it tramples the space reserved by uint32_t
data[4] in struct pbo_spec_async_data.  This ends up setting
created in pbo_spec_async_data to true leading to a SIGSEGV crash when
a null nir field is accessed.

Fixed by changing uint16_t to uint8_t alignment.  It's a 2-bit field
anyway so those extra bits aren't needed, that then matches with the
types before and after in the union, and pbo_data is 16 bytes in MSVC
and Clang compiled programs too.

See https://github.com/Juice-Labs/juice-hp/issues/1679.

+++ src/mesa/state_tracker/st_pbo_compute.c
@@ -182,7 +182,7 @@ struct pbo_data {
              uint8_t r11g11b10_or_sint : 1;
              uint8_t r9g9b9e5 : 1;
              uint8_t swap : 1;
-             uint16_t alignment : 2;
+             uint8_t alignment : 2;
              uint8_t dst_bit_size : 2; //8, 16, 32, 64
           };
